Trying the floar operater

+++ Starting/Arithmetic.java
@@ -3,7 +3,7 @@ public class Arithmetic {
     public static void main(String args[])
     {
         int a = 14, b = 5;
-        int c = a / b;
+        float c =(float)a / b;
         int r = a % b;
 
         System.out.println(c);
